+++ save.c
@@ -455,7 +455,7 @@ void fwrite_char( CHAR_DATA * ch, FILE * fp )
       if( ch->pcdata->pdeaths )
          fprintf( fp, "PDeaths      %d\n", ch->pcdata->pdeaths );
       if( get_timer( ch, TIMER_PKILLED ) && ( get_timer( ch, TIMER_PKILLED ) > 0 ) )
-         fprintf( fp, "PTimer       %d\n", get_timer( ch, TIMER_PKILLED ) );
+         fprintf( fp, "PTimer       %d\n", (int)get_timer( ch, TIMER_PKILLED ) );
       fprintf( fp, "MKills       %d\n", ch->pcdata->mkills );
       fprintf( fp, "MDeaths      %d\n", ch->pcdata->mdeaths );
       if( ch->pcdata->illegal_pk )
